+++ backend/taramail/testing/managers.py
@@ -8,25 +8,20 @@ from taramail.alias import AliasManager
 from taramail.dkim import DKIMManager
 from taramail.domain import DomainManager
 from taramail.mailbox import MailboxManager
+from taramail.password import PasswordPolicyManager
 from taramail.sogo import Sogo
-
-
-@pytest.fixture
-def sogo(db_session, memcached_store):
-    """Sogo fixture."""
-    return Sogo(db_session, memcached_store)
-
-
-@pytest.fixture
-def dkim_manager(redis_store):
-    """DKIM manager fixture."""
-    return DKIMManager(redis_store)
 
 
 @pytest.fixture
 def alias_manager(db_session):
     """Alias manager fixture."""
     return AliasManager(db_session)
+
+
+@pytest.fixture
+def dkim_manager(redis_store):
+    """DKIM manager fixture."""
+    return DKIMManager(redis_store)
 
 
 @pytest.fixture
@@ -37,6 +32,22 @@ def domain_manager(db_session, redis_store):
 
 
 @pytest.fixture
-def mailbox_manager(db_session, redis_store, sogo):
+def mailbox_manager(db_session, redis_store, password_policy_manager, sogo):
     """Mailbox manager fixture."""
-    return MailboxManager(db_session, redis_store, sogo)
+    return MailboxManager(db_session, redis_store, password_policy_manager, sogo)
+
+
+@pytest.fixture
+def password_policy_manager(redis_store):
+    """Password policy manager fixture."""
+    manager = PasswordPolicyManager(redis_store)
+    try:
+        yield manager
+    finally:
+        manager.reset_policy()
+
+
+@pytest.fixture
+def sogo(db_session, memcached_store):
+    """Sogo fixture."""
+    return Sogo(db_session, memcached_store)